code monitors: extract email action to its own component
