+++ client/web/src/enterprise/code-monitoring/CreateCodeMonitorPage.test.tsx
@@ -35,11 +35,11 @@ describe('CreateCodeMonitorPage', () => {
     }
     let clock: sinon.SinonFakeTimers
 
-    beforeAll(() => {
+    beforeEach(() => {
         clock = sinon.useFakeTimers()
     })
 
-    afterAll(() => {
+    afterEach(() => {
         clock.restore()
     })
 
@@ -68,6 +68,10 @@ describe('CreateCodeMonitorPage', () => {
         userEvent.click(screen.getByTestId('form-action-toggle-email-notification'))
 
         userEvent.click(screen.getByTestId('submit-action'))
+
+        act(() => {
+            clock.tick(600)
+        })
 
         userEvent.click(screen.getByTestId('submit-monitor'))
 
@@ -103,6 +107,10 @@ describe('CreateCodeMonitorPage', () => {
         userEvent.click(screen.getByTestId('form-action-toggle-email-notification'))
         userEvent.click(screen.getByTestId('submit-action'))
 
+        act(() => {
+            clock.tick(600)
+        })
+
         // Pressing enter calls createCodeMonitor when all sections are complete
         userEvent.click(screen.getByTestId('submit-monitor'))
 
@@ -112,6 +120,6 @@ describe('CreateCodeMonitorPage', () => {
     test('Actions area button is disabled while trigger is incomplete', () => {
         render(<CreateCodeMonitorPage {...props} />)
         const actionButton = screen.getByTestId('form-action-toggle-email-notification')
-        expect(actionButton).toBeDisabled()
+        expect(actionButton).toHaveClass('disabled')
     })
 })
