+++ client/web/src/enterprise/code-monitoring/backend.ts
@@ -182,6 +182,7 @@ export const fetchCodeMonitor = (id: string): Observable<FetchCodeMonitorResult>
                     enabled
                     actions {
                         nodes {
+                            __typename
                             ... on MonitorEmail {
                                 id
                                 recipients {
